AbstractQuantityValue (setValue/getValue have different param/return types)
