+++ models/DataObject/Data/InputQuantityValue.php
@@ -16,18 +16,33 @@
 
 namespace Pimcore\Model\DataObject\Data;
 
-use Pimcore\Model\DataObject\OwnerAwareFieldInterface;
 use Pimcore\Model\DataObject\QuantityValue\Unit;
-use Pimcore\Model\DataObject\Traits\OwnerAwareFieldTrait;
 
-class InputQuantityValue extends QuantityValue implements OwnerAwareFieldInterface
+class InputQuantityValue extends AbstractQuantityValue
 {
-    use OwnerAwareFieldTrait;
-
     /**
-     * @var null|string
+     * @var string|null
      */
     protected $value;
+
+    /**
+     * @param string|null $value
+     * @param Unit|string|null $unit
+     */
+    public function __construct($value = null, $unit = null)
+    {
+        $this->value = $value;
+        parent::__construct($unit);
+    }
+
+    /**
+     * @param string $value
+     */
+    public function setValue($value)
+    {
+        $this->value = $value;
+        $this->markMeDirty();
+    }
 
     /**
      * @return string
